# Fix performance issue on the status page