Fixes https://github.com/vitessio/arewefastyet/issues/521

```
[GIN] 2024/04/02 - 14:53:00 | 200 |  788.178958ms |             ::1 | GET      "/api/recent"
[GIN] 2024/04/02 - 14:53:01 | 200 |   94.367042ms |             ::1 | GET      "/api/recent"
[GIN] 2024/04/02 - 14:53:02 | 200 |   93.952292ms |             ::1 | GET      "/api/recent"
[GIN] 2024/04/02 - 14:53:03 | 200 |  182.713041ms |             ::1 | GET      "/api/recent"
[GIN] 2024/04/02 - 14:53:04 | 200 |  179.287459ms |             ::1 | GET      "/api/recent"
```